Cambiar tipo de retorno del GetScore de float -- a --> int

+++ Unity_project/Mycrobiology_project/Assets/Puzzles/Commons/IPuzzleManager/IPuzzleManager.cs
@@ -7,7 +7,7 @@ public interface IPuzzleManager
 
     public void CompletePuzzle();
 
-    public float GetScore();
+    public int GetScore();
 
     public PerformanceResult GetPerformanceResult();
 }
